+++ Photos-QR-Scanner/PhotoMetadataApp.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+
+struct PhotoMetadataApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
